Rename tabby collections for consistency
The sfb tabby specification (1) said that the value provided as the
dataset name "should be suitable for a directory/folder name. Spaces
and special characters should be avoided" but this was not enforced,
so the collection (i.e. folder) names are only loosely coupled to
the name field.

This commit introduces semi-consistency in the tabby collection
names, so that the following applies:

  - if the provided name has no spaces, it is used verbatim
  - if the name has spaces, they are replaced with "-"
  - there are some exceptions (names overly generic, or agreed later)

I have been avoiding changing the name stored in the tabby file
because it is used to generate the dataset ID.

(1)  https://rdm.sfb1451.de/data-catalog-submission/